fix (#31): add type string and null in return type of array response and add parse the offset and totalPages value to int type

+++ src/Utils/Pagination.php
@@ -16,7 +16,7 @@ abstract class Pagination {
      * @param int $countItems
      * @param UrlGeneratorInterface $urlGenerator
      *
-     * @return array<string, int>
+     * @return array<string, int|string|null>
      */
     public static function getValueParamsToPagination(
         Request $request,
@@ -62,8 +62,8 @@ abstract class Pagination {
             $paramsToUrlGenerate['query']['limit'] = $limit;
         }
 
-        $offset = ($page - 1) * $limit;
-        $totalPages = ceil($countItems / $limit);
+        $offset = (int) (($page - 1) * $limit);
+        $totalPages = (int) ceil($countItems / $limit);
 
         if ($page > $totalPages) {
             throw new NotFoundHttpException("Page not found");
